+++ CHANGELOG.md
@@ -44,6 +44,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix Spelling
 
+### Perf
+
+- use dc_ref instead of $ref in serialization to avoid overhead during CRUD operation on platform
+
 
 ## 0.13.2
 
@@ -53,6 +57,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fix
 
+- Workflow state/run to_dict fix on adding references
+- Handle serialization pointers of non-standalone objects
+- hash fix: calling hash instead of data hash in eq.
+
 ## 0.13.1
 
 ### Fix
@@ -61,7 +69,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - WorkflowRun settings now sets the right method to call for workflow display
 
 
-## 0.13.0
+## 0.13.0 [02/14/2023]
+
 
 ### Chore
 
@@ -98,7 +107,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fix str of Dataset
 
-## 0.12.0
+## 0.12.0 [released 01/20/2023]
 
 
 ### Changed
@@ -150,7 +159,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - More docs
 
 
-## 0.11.0
+## 0.11.0 [released 12/19/2022]
 
 ### Fixed